# openapi/devflow-lengkap.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevFlow — API Lengkap",
    "version": "1.0",
    "description": "Gabungan spesifikasi kelima service backend DevFlow, seluruhnya di satu berkas.\n\nDihasilkan dari controller lewat springdoc, lalu digabung — bukan ditulis tangan, jadi isinya mengikuti kode.\n\nSetiap operasi diberi tag nama service asalnya. Ambil token lewat POST /api/v1/auth/login, lalu pakai skema bearerAuth."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "api-gateway"
    }
  ],
...
  "paths": {
    "/oauth2/google/callback": {
      "post": {
        "tags": [
          "auth-service"
        ],
        "operationId": "googleCallback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/oauth2/github/connect/callback": {
      "post": {
        "tags": [
          "auth-service"
        ],
        "operationId": "githubConnectCallback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/oauth2/github/callback": {
      "post": {
        "tags": [
          "auth-service"
        ],
        "operationId": "githubCallback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/internal": {
      "post": {
        "tags": [
          "auth-service"
        ],
        "operationId": "create",
        "parameters": [
          {
            "name": "X-Internal-Token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media": {
      "post": {
        "tags": [
          "auth-service"
        ],
        "operationId": "upload",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "AVATAR",
                "WORKSPACE_LOGO"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "tags": [
          "auth-service"
        ],
        "operationId": "markAsRead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "auth-service"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageNotificationDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/media/{id}": {
      "get": {
        "tags": [
          "auth-service"
        ],
        "operationId": "serve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{id}": {
      "delete": {
        "tags": [
          "auth-service"
        ],
        "operationId": "archive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CallbackRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 512,
            "minLength": 0
          },
          "state": {
            "type": "string",
            "maxLength": 512,
            "minLength": 0
          }
        },
        "required": [
          "code"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "profileComplete": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "DELETED"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "globalRole": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN"
            ]
          },
          "identities": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GOOGLE",
                "GITHUB"
              ]
            }
          }
        }
      },
      "CreateNotificationRequest": {
        "type": "object",
        "properties": {
          "recipientEmail": {
            "type": "string"
          },
          "recipientUserId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "content": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 0
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {}
          },
          "priority": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          },
          "actionUrl": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {}
          },
          "status": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "actionUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PageNotificationDto": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Access token dari /api/v1/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}